Bug 666248: Visual C++ compilation support
This enables libcroco to be compiled with Visual C++ 2008/2010, along
with its test programs.  A libcroco.symbols is included to export the
symbols from the DLL that is built in the process, and the projects
for the main libcroco DLL is done in a way where their source file
listings are filled in during "make dist" as they could be changed from
time to time so to simplify maintenanace work.  Versioning stuff are
also done via autotools as far as possible also.

The .sln files have no simple way to include the LGPL 2.1+ terms without
causing trouble, so the text for them is included as follows, for
references.
------
This file is part of The Croco Library
This program is free software; you can redistribute it and/or
modify it under the terms of version 2.1 of the GNU Lesser General Public
License as published by the Free Software Foundation.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU Lesser General Public License
along with this program; if not, write to the Free Software
Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
USA

Author: Fan, Chun-wei
See COPYRIGHTS file for copyright information.
------