Visual Studio Build Files Cleanup and Fixes
Fix up the property sheets, as it was "installing" the libcroco stuff one
level down from the desired folder, and $(top_srcdir) was missed from the
include directories list for the VS2010 property sheets.  Also unify items
so that we can reduce duplicate items in there.

Build all projects with the MultiByte character set setting to be
consistent with the rest of the GLib stack.