MSVC Builds: Split up the Property Sheets
Update the Visual Studio Projects so that the property sheets are split by
the functions required, so to simplify their maintenance.  The projects can
then import the property sheets that suit their needs.  This is done like
the rest of the GTK+ stack, so it also makes it easier to support newer
versions of Visual Studio.

As the original main property sheet is generated during configure time (and
disted), update configure.ac as well, as the name of the template file
changed in the process.