MSVC Builds: Add Simple Support for MSVC 2012/2013
As the Visual Studio 2012/2013 project file format is largely the same as
the Visual Studio 2010 ones, we can add support for them with relative ease
by copying the 2010 files and updating them as necessary with autotools
scripts, so to keep them up-to-date, all that is needed is that the Visual
Studio 2010 projects are kept up to date.

This might change if we do support the stack under the Windows RT APIs, but
this will do the job for now.