Visual Studio builds: Support Visual Studio 2017
This updates the autotools scripts so that we can support Visual Studio 2017
by copying the 2010 projects and update items in there as needed.

Note that the format of the toolset version string changed for Visual Studio
2017, so allow a custom toolset version string to be passed in and used,
otherwise the platform toolset string is generated as before.

Note also that Visual Studio 2017 aims to be compatible with 2015 on the
CRT level, so one should be able to use 2017-built binaries with 2015-built
binaries without problems.